[REFACTOR] 어노테이션 사용해서 의존성 자동 주입되도록 수정 (@Autowired)
